refactor(screen-capture): flip to plugin-sdk + park the Apple ScreenCaptureKit module (#1263)
screen-capture is the last of the eight facade-linking packages to move onto the
engine-free plugin-authoring SDK. attempt-2 reverted it to the `streamlib` facade
as a "choice"; the facade is the migration artifact being deleted, so nothing
stays facade-linked. Swap the runtime dep `streamlib` -> `streamlib-plugin-sdk`
and retarget `streamlib::sdk::*` -> `streamlib_plugin_sdk::sdk::*`.

Its sole processor is the Apple ScreenCaptureKit capture, which names an
engine-free Apple RHI / surface-pool surface (`sdk::rhi::{PixelBuffer,
PixelBufferRef, PixelFormat}` + surface-pool acquire/check-in) that the plugin
SDK carries only under `cfg(linux)`. Following the mp4 precedent, the module
moves to `src/_apple_impl_pending_/` behind a never-compile `#[cfg(any())]` gate
so it is not built against the missing surface on any target; the package now
ships no live processor and compiles clean on Linux and aarch64-apple-darwin.
The in-code marker names the unpark condition (plugin SDK gains an engine-free
Apple RHI / surface-pool surface) and the Cargo.toml Apple interop deps to
restore alongside it.

check-boundaries drops screen-capture from the packages-facade-dep allowlist;
all eight flipped packages are now off it, leaving only the two permanent
host-by-design entries plus the shrinking primitive-gated backlog.
