+++ packages/screen-capture/src/_apple_impl_pending_/corevideo_ffi.rs
